refactor(canon): split package tsconfig checks
